Module rustc_resolve::build_reduced_graph

After we obtain a fresh AST fragment from a macro, code in this module helps to integrate that fragment into the module structures that are already partially built.

Items from the fragment are placed into modules, unexpanded macros in the fragment are visited and registered. Imports are also considered items and placed into modules here, but not resolved yet.

Macros

method 🔒

Structs

BuildReducedGraphVisitor 🔒

Type Aliases

Res 🔒