Struct rustc_resolve::build_reduced_graph::BuildReducedGraphVisitor

struct BuildReducedGraphVisitor<'a, 'b> {
    r: &'b mut Resolver<'a>,
    parent_scope: ParentScope<'a>,
}

Fields
r: &'b mut Resolver<'a>parent_scope: ParentScope<'a>
Implementations

impl<'a, 'b> BuildReducedGraphVisitor<'a, 'b>

fn resolve_visibility(&mut self, vis: &Visibility) -> Visibility

fn try_resolve_visibility<'ast>(
&mut self,
vis: &'ast Visibility,
finalize: bool
) -> Result<Visibility, VisResolutionError<'ast>>

fn insert_field_names_local(&mut self, def_id: DefId, vdata: &VariantData)

fn insert_field_names_extern(&mut self, def_id: DefId)

fn block_needs_anonymous_module(&mut self, block: &Block) -> bool

fn add_import(
&mut self,
module_path: Vec<Segment>,
kind: ImportKind<'a>,
span: Span,
item: &Item,
root_span: Span,
root_id: NodeId,
vis: Visibility
)

fn build_reduced_graph_for_use_tree(
&mut self,
use_tree: &UseTree,
id: NodeId,
parent_prefix: &[Segment],
nested: bool,
item: &Item,
vis: Visibility,
root_span: Span
)

fn build_reduced_graph_for_item(&mut self, item: &'b Item)

Constructs the reduced graph for one item.

fn build_reduced_graph_for_extern_crate(
&mut self,
orig_name: Option<Symbol>,
item: &Item,
local_def_id: LocalDefId,
vis: Visibility,
parent: &'a ModuleData<'a>
)

fn build_reduced_graph_for_foreign_item(&mut self, item: &ForeignItem)

Constructs the reduced graph for one foreign item.

fn build_reduced_graph_for_block(&mut self, block: &Block)

fn build_reduced_graph_for_external_crate_res(&mut self, child: ModChild)

Builds the reduced graph for a single item in an external crate.

fn add_macro_use_binding(
&mut self,
name: Symbol,
binding: &'a NameBinding<'a>,
span: Span,
allow_shadowing: bool
)

fn process_macro_use_imports(
&mut self,
item: &Item,
module: &'a ModuleData<'a>
) -> bool

Returns true if we should consider the underlying extern crate to be used.

fn contains_macro_use(&mut self, attrs: &[Attribute]) -> bool

Returns true if this attribute list contains macro_use.

fn visit_invoc(&mut self, id: NodeId) -> LocalExpnId

fn visit_invoc_in_module(
&mut self,
id: NodeId
) -> Interned<'a, Cell<MacroRulesScope<'a>>>

Visit invocation in context in which it can emit a named item (possibly macro_rules) directly into its parent scope’s module.

fn proc_macro_stub(&self, item: &Item) -> Option<(MacroKind, Ident, Span)>

fn insert_unused_macro(
&mut self,
ident: Ident,
def_id: LocalDefId,
node_id: NodeId,
rule_spans: &[(usize, Span)]
)

fn define_macro(
&mut self,
item: &Item
) -> Interned<'a, Cell<MacroRulesScope<'a>>>