Function rustc_resolve::create_resolver

pub fn create_resolver<'a, 'tcx>(session: &'a Session, ast_map: &'a Map<'tcx>, krate: &'a Crate, make_glob_map: MakeGlobMap, arenas: &'a ResolverArenas<'a>, callback: Option<Box<Fn(Node, &mut bool) -> bool>>) -> Resolver<'a, 'tcx>

Unstable (rustc_private)

Builds a name resolution walker to be used within this module, or used externally, with an optional callback function.

The callback takes a &mut bool which allows callbacks to end a walk when set to true, passing through the rest of the walk, while preserving the ribs + current module. This allows resolve_path calls to be made with the correct scope info. The node in the callback corresponds to the current node in the walk.