Module rustc_resolve::late::diagnostics

Structs

ElisionFnParameter 🔒

Description of the lifetimes appearing in a function parameter. This is used to provide a literal explanation to the elision failure.

MissingLifetime 🔒

Description of an elided lifetime.

Enums

AssocSuggestion 🔒

A field or associated item from self type suggested in case of resolution failure.

LifetimeElisionCandidate 🔒

Description of lifetimes that appear as candidates for elision. This is used to suggest introducing an explicit lifetime.

MissingLifetimeKind 🔒

Functions

import_candidate_to_enum_paths 🔒

Gets the stringified path for an enum from an ImportSuggestion for an enum variant.

is_self_type 🔒

is_self_value 🔒

signal_label_shadowing

Shadowing involving a label is only a warning for historical reasons.

signal_lifetime_shadowing

Report lifetime/lifetime shadowing as an error.

Type Definitions

Res 🔒