Enum rustc_resolve::late::LifetimeRibKind

enum LifetimeRibKind {
    Generics {
        binder: NodeId,
        span: Span,
        kind: LifetimeBinderKind,
    },
    AnonymousCreateParameter {
        binder: NodeId,
        report_in_path: bool,
    },
    Elided(LifetimeRes),
    AnonymousReportError,
    AnonymousWarnToStatic(NodeId),
    ElisionFailure,
    ConstParamTy,
    ConcreteAnonConst(NoConstantGenericsReason),
    Item,
}

Variants

Generics

Fields

binder: NodeId

span: Span

kind: LifetimeBinderKind

This rib declares generic parameters. Only for this kind the LifetimeRib::bindings field can be non-empty.

AnonymousCreateParameter

Fields

binder: NodeId

report_in_path: bool

Create a new anonymous lifetime parameter and reference it.

If report_in_path, report an error when encountering lifetime elision in a path:

struct Foo<'a> { x: &'a () }
async fn foo(x: Foo) {}

Note: the error should not trigger when the elided lifetime is in a pattern or expression-position path:

struct Foo<'a> { x: &'a () }
async fn foo(Foo { x: _ }: Foo<'_>) {}

Elided(LifetimeRes)

Replace all anonymous lifetimes by provided lifetime.

AnonymousReportError

Give a hard error when either & or '_ is written. Used to rule out things like where T: Foo<'_>. Does not imply an error on default object bounds (e.g., Box<dyn Foo>).

AnonymousWarnToStatic(NodeId)

Resolves elided lifetimes to 'static, but gives a warning that this behavior is a bug and will be reverted soon.

ElisionFailure

Signal we cannot find which should be the anonymous lifetime.

ConstParamTy

This rib forbids usage of generic parameters inside of const parameter types.

While this is desirable to support eventually, it is difficult to do and so is currently forbidden. See rust-lang/project-const-generics#28 for more info.

ConcreteAnonConst(NoConstantGenericsReason)

Usage of generic parameters is forbidden in various positions for anon consts:

const arguments when generic_const_exprs is not enabled

enum discriminant values

This rib emits an error when a lifetime would resolve to a lifetime parameter.

Item

This rib acts as a barrier to forbid reference to lifetimes of a parent item.