Enum rustc_resolve::late::lifetimes::Scope

enum Scope<'a> {
    Binder {
        lifetimes: FxIndexMap<LocalDefId, Region>,
        scope_type: BinderScopeType,
        hir_id: HirId,
        s: &'a Scope<'a>,
        where_bound_origin: Option<PredicateOrigin>,
    },
    Body {
        id: BodyId,
        s: &'a Scope<'a>,
    },
    Elision {
        s: &'a Scope<'a>,
    },
    ObjectLifetimeDefault {
        lifetime: Option<Region>,
        s: &'a Scope<'a>,
    },
    Supertrait {
        lifetimes: Vec<BoundVariableKind>,
        s: &'a Scope<'a>,
    },
    TraitRefBoundary {
        s: &'a Scope<'a>,
    },
    Root,
}

Variants

Binder

Fields

lifetimes: FxIndexMap<LocalDefId, Region>

We use an IndexMap here because we want these lifetimes in order for diagnostics.

scope_type: BinderScopeType

hir_id: HirId

The late bound vars for a given item are stored by HirId to be queried later. However, if we enter an elision scope, we have to later append the elided bound vars to the list and need to know what to append to.

s: &'a Scope<'a>

where_bound_origin: Option<PredicateOrigin>

If this binder comes from a where clause, specify how it was created. This is used to diagnose inaccessible lifetimes in APIT:

fn foo(x: impl for<'a> Trait<'a, Assoc = impl Copy + 'a>) {}

Declares lifetimes, and each can be early-bound or late-bound. The DebruijnIndex of late-bound lifetimes starts at 1 and it should be shifted by the number of Binders in between the declaration Binder and the location it’s referenced from.

Body

Fields

id: BodyId

s: &'a Scope<'a>

Lifetimes introduced by a fn are scoped to the call-site for that fn, if this is a fn body, otherwise the original definitions are used. Unspecified lifetimes are inferred, unless an elision scope is nested, e.g., (&T, fn(&T) -> &T); becomes (&'_ T, for<'a> fn(&'a T) -> &'a T).

Elision

Fields

s: &'a Scope<'a>

A scope which either determines unspecified lifetimes or errors on them (e.g., due to ambiguity).

ObjectLifetimeDefault

Fields

lifetime: Option<Region>

s: &'a Scope<'a>

Use a specific lifetime (if Some) or leave it unset (to be inferred in a function body or potentially error outside one), for the default choice of lifetime in a trait object type.

Supertrait

Fields

lifetimes: Vec<BoundVariableKind>

s: &'a Scope<'a>

When we have nested trait refs, we concatenate late bound vars for inner trait refs from outer ones. But we also need to include any HRTB lifetimes encountered when identifying the trait that an associated type is declared on.

TraitRefBoundary

Fields

s: &'a Scope<'a>

Root