Struct rustc_resolve::late::DiagMetadata

struct DiagMetadata<'ast> {

    current_trait_assoc_items: Option<&'ast [P<AssocItem>]>,
    current_self_type: Option<Ty>,
    current_self_item: Option<NodeId>,
    current_item: Option<&'ast Item>,
    currently_processing_generic_args: bool,
    current_function: Option<(FnKind<'ast>, Span)>,
    unused_labels: FxHashMap<NodeId, Span>,
    current_block_could_be_bare_struct_literal: Option<Span>,
    current_let_binding: Option<(Span, Option<Span>, Option<Span>)>,
    current_pat: Option<&'ast Pat>,
    in_if_condition: Option<&'ast Expr>,
    in_assignment: Option<&'ast Expr>,
    is_assign_rhs: bool,
    in_range: Option<(&'ast Expr, &'ast Expr)>,
    current_trait_object: Option<&'ast [GenericBound]>,
    current_where_predicate: Option<&'ast WherePredicate>,
    current_type_path: Option<&'ast Ty>,
    current_impl_items: Option<&'ast [P<AssocItem>]>,
    currently_processing_impl_trait: Option<(TraitRef, Ty)>,
    current_elision_failures: Vec<MissingLifetime>,

}

Fields
current_trait_assoc_items: Option<&'ast [P<AssocItem>]>

The current trait’s associated items’ ident, used for diagnostic suggestions.

current_self_type: Option<Ty>

The current self type if inside an impl (used for better errors).

current_self_item: Option<NodeId>

The current self item if inside an ADT (used for better errors).

current_item: Option<&'ast Item>

The current trait (used to suggest).

currently_processing_generic_args: bool

When processing generic arguments and encountering an unresolved ident not found, suggest introducing a type or const param depending on the context.

current_function: Option<(FnKind<'ast>, Span)>

The current enclosing (non-closure) function (used for better errors).

unused_labels: FxHashMap<NodeId, Span>

A list of labels as of yet unused. Labels will be removed from this map when they are used (in a break or continue statement)

current_block_could_be_bare_struct_literal: Option<Span>

Only used for better errors on let x = { foo: bar };. In the case of a parse error with let x = { foo: bar, };, this isn’t needed, it’s only needed for cases where this parses as a correct type ascription.

current_let_binding: Option<(Span, Option<Span>, Option<Span>)>

Only used for better errors on let <pat>: <expr, not type>;.

current_pat: Option<&'ast Pat>in_if_condition: Option<&'ast Expr>

Used to detect possible if let written without let and to provide structured suggestion.

in_assignment: Option<&'ast Expr>

Used to detect possible new binding written without let and to provide structured suggestion.

is_assign_rhs: boolin_range: Option<(&'ast Expr, &'ast Expr)>

Used to detect possible . -> .. typo when calling methods.

current_trait_object: Option<&'ast [GenericBound]>

If we are currently in a trait object definition. Used to point at the bounds when encountering a struct or enum.

current_where_predicate: Option<&'ast WherePredicate>

Given where <T as Bar>::Baz: String, suggest where T: Bar<Baz = String>.

current_type_path: Option<&'ast Ty>current_impl_items: Option<&'ast [P<AssocItem>]>

The current impl items (used to suggest).

currently_processing_impl_trait: Option<(TraitRef, Ty)>

When processing impl trait

current_elision_failures: Vec<MissingLifetime>

Accumulate the errors due to missed lifetime elision, and report them all at once for each function.