Struct rustc_resolve::late::LateResolutionVisitor

struct LateResolutionVisitor<'a, 'b, 'ast> {
    r: &'b mut Resolver<'a>,
    parent_scope: ParentScope<'a>,
    ribs: PerNS<Vec<Rib<'a>>>,
    label_ribs: Vec<Rib<'a, NodeId>>,
    lifetime_ribs: Vec<LifetimeRib>,
    lifetime_elision_candidates: Option<FxIndexMap<LifetimeRes, LifetimeElisionCandidate>>,
    current_trait_ref: Option<(&'a ModuleData<'a>, TraitRef)>,
    diagnostic_metadata: Box<DiagnosticMetadata<'ast>>,
    in_func_body: bool,
    lifetime_uses: FxHashMap<LocalDefId, LifetimeUseSet>,
}

Fields
r: &'b mut Resolver<'a>parent_scope: ParentScope<'a>

The module that represents the current item scope.

ribs: PerNS<Vec<Rib<'a>>>

The current set of local scopes for types and values. FIXME #4948: Reuse ribs to avoid allocation.

label_ribs: Vec<Rib<'a, NodeId>>

The current set of local scopes, for labels.

lifetime_ribs: Vec<LifetimeRib>

The current set of local scopes for lifetimes.

lifetime_elision_candidates: Option<FxIndexMap<LifetimeRes, LifetimeElisionCandidate>>

We are looking for lifetimes in an elision context. The set contains all the resolutions that we encountered so far. They will be used to determine the correct lifetime for the fn return type. The LifetimeElisionCandidate is used for diagnostics, to suggest introducing named lifetimes.

current_trait_ref: Option<(&'a ModuleData<'a>, TraitRef)>

The trait that the current context can refer to.

diagnostic_metadata: Box<DiagnosticMetadata<'ast>>

Fields used to add information to diagnostic errors.

in_func_body: bool

State used to know whether to ignore resolution errors for function bodies.

In particular, rustdoc uses this to avoid giving errors for cfg() items. In most cases this will be None, in which case errors will always be reported. If it is true, then it will be updated when entering a nested function or trait body.

lifetime_uses: FxHashMap<LocalDefId, LifetimeUseSet>

Count the number of places a lifetime is used.

Implementations

impl<'a: 'ast, 'ast> LateResolutionVisitor<'a, '_, 'ast>

fn def_span(&self, def_id: DefId) -> Option<Span>

pub(crate) fn smart_resolve_report_errors(
&mut self,
path: &[Segment],
span: Span,
source: PathSource<'_>,
res: Option<Res<NodeId>>
) -> (DiagnosticBuilder<'a, ErrorGuaranteed>, Vec<ImportSuggestion>)

Handles error reporting for smart_resolve_path_fragment function. Creates base error and amends it with one short label and possibly some longer helps/notes.

fn detect_assoct_type_constraint_meant_as_path(
&self,
base_span: Span,
err: &mut Diagnostic
)

fn suggest_swapping_misplaced_self_ty_and_trait(
&mut self,
err: &mut Diagnostic,
source: PathSource<'_>,
res: Option<Res<NodeId>>,
span: Span
)

fn get_single_associated_item(
&mut self,
path: &[Segment],
source: &PathSource<'_>,
filter_fn: &impl Fn(Res<NodeId>) -> bool
) -> Option<TypoSuggestion>

fn restrict_assoc_type_in_where_clause(
&mut self,
span: Span,
err: &mut Diagnostic
) -> bool

Given where <T as Bar>::Baz: String, suggest where T: Bar<Baz = String>.

fn call_has_self_arg(
&self,
source: PathSource<'_>
) -> Option<(Span, Option<Span>)>

Check if the source is call expression and the first argument is self. If true, return the span of whole call and the span for all arguments expect the first one (self).

fn followed_by_brace(&self, span: Span) -> (bool, Option<Span>)

fn smart_resolve_context_dependent_help(
&mut self,
err: &mut Diagnostic,
span: Span,
source: PathSource<'_>,
res: Res<NodeId>,
path_str: &str,
fallback_label: &str
) -> bool

Provides context-dependent help for errors reported by the smart_resolve_path_fragment function. Returns true if able to provide context-dependent help.

pub(crate) fn find_similarly_named_assoc_item(
&mut self,
ident: Symbol,
kind: &AssocItemKind
) -> Option<Symbol>

Given the target ident and kind, search for the similarly named associated item in self.current_trait_ref.

fn lookup_assoc_candidate<FilterFn>(
&mut self,
ident: Ident,
ns: Namespace,
filter_fn: FilterFn
) -> Option<AssocSuggestion>where
FilterFn: Fn(Res<NodeId>) -> bool,

fn lookup_typo_candidate(
&mut self,
path: &[Segment],
ns: Namespace,
filter_fn: &impl Fn(Res<NodeId>) -> bool
) -> Option<TypoSuggestion>

fn likely_rust_type(path: &[Segment]) -> Option<Symbol>

fn get_colon_suggestion_span(&self, start: Span) -> Span

Only used in a specific case of type ascription suggestions

fn type_ascription_suggestion(
&self,
err: &mut Diagnostic,
base_span: Span
) -> bool

fn find_module(
&mut self,
def_id: DefId
) -> Option<(&'a ModuleData<'a>, ImportSuggestion)>

fn collect_enum_ctors(
&mut self,
def_id: DefId
) -> Option<Vec<(Path, DefId, CtorKind)>>

fn suggest_using_enum_variant(
&mut self,
err: &mut Diagnostic,
source: PathSource<'_>,
def_id: DefId,
span: Span
)

Adds a suggestion for using an enum’s variant when an enum is used instead.

pub(crate) fn report_missing_type_error(
&self,
path: &[Segment]
) -> Option<(Span, &'static str, String, Applicability)>

pub(crate) fn suggestion_for_label_in_rib(
&self,
rib_index: usize,
label: Ident
) -> Option<(Ident, bool)>

Given the target label, search the rib_indexth label rib for similarly named labels, optionally returning the closest match and whether it is reachable.

pub(crate) fn maybe_report_lifetime_uses(
&mut self,
generics_span: Span,
params: &[GenericParam]
)

pub(crate) fn emit_undeclared_lifetime_error(
&self,
lifetime_ref: &Lifetime,
outer_lifetime_ref: Option<Ident>
)

fn suggest_introducing_lifetime(
&self,
err: &mut Diagnostic,
name: Option<&str>,
suggest: impl Fn(&mut Diagnostic, bool, Span, &str, String) -> bool
)

pub(crate) fn emit_non_static_lt_in_const_generic_error(
&self,
lifetime_ref: &Lifetime
)

pub(crate) fn maybe_emit_forbidden_non_static_lifetime_error(
&self,
lifetime_ref: &Lifetime
)

Non-static lifetimes are prohibited in anonymous constants under min_const_generics. This function will emit an error if generic_const_exprs is not enabled, the body identified by body_id is an anonymous constant and lifetime_ref is non-static.

pub(crate) fn report_missing_lifetime_specifiers(
&mut self,
lifetime_refs: Vec<MissingLifetime>,
function_param_lifetimes: Option<(Vec<MissingLifetime>, Vec<ElisionFnParameter>)>
) -> ErrorGuaranteed

fn add_missing_lifetime_specifiers_label(
&mut self,
err: &mut Diagnostic,
lifetime_refs: Vec<MissingLifetime>,
function_param_lifetimes: Option<(Vec<MissingLifetime>, Vec<ElisionFnParameter>)>
)

impl<'a: 'ast, 'b, 'ast> LateResolutionVisitor<'a, 'b, 'ast>

fn new(resolver: &'b mut Resolver<'a>) -> LateResolutionVisitor<'a, 'b, 'ast>

fn maybe_resolve_ident_in_lexical_scope(
&mut self,
ident: Ident,
ns: Namespace
) -> Option<LexicalScopeBinding<'a>>

fn resolve_ident_in_lexical_scope(
&mut self,
ident: Ident,
ns: Namespace,
finalize: Option<Finalize>,
ignore_binding: Option<&'a NameBinding<'a>>
) -> Option<LexicalScopeBinding<'a>>

fn resolve_path(
&mut self,
path: &[Segment],
opt_ns: Option<Namespace>,
finalize: Option<Finalize>
) -> PathResult<'a>

fn with_rib<T>(
&mut self,
ns: Namespace,
kind: RibKind<'a>,
work: impl FnOnce(&mut Self) -> T
) -> T

Do some work within a new innermost rib of the given kind in the given namespace (ns).

fn with_scope<T>(&mut self, id: NodeId, f: impl FnOnce(&mut Self) -> T) -> T

fn visit_generic_params(
&mut self,
params: &'ast [GenericParam],
add_self_upper: bool
)

fn with_lifetime_rib<T>(
&mut self,
kind: LifetimeRibKind,
work: impl FnOnce(&mut Self) -> T
) -> T

fn resolve_lifetime(&mut self, lifetime: &'ast Lifetime, use_ctxt: LifetimeCtxt)

fn resolve_anonymous_lifetime(&mut self, lifetime: &Lifetime, elided: bool)

fn resolve_elided_lifetime(&mut self, anchor_id: NodeId, span: Span)

fn create_fresh_lifetime(
&mut self,
id: NodeId,
ident: Ident,
binder: NodeId
) -> LifetimeRes

fn resolve_elided_lifetimes_in_path(
&mut self,
path_id: NodeId,
partial_res: PartialRes,
path: &[Segment],
source: PathSource<'_>,
path_span: Span
)

fn record_lifetime_res(
&mut self,
id: NodeId,
res: LifetimeRes,
candidate: LifetimeElisionCandidate
)

fn record_lifetime_param(&mut self, id: NodeId, res: LifetimeRes)

fn resolve_fn_signature(
&mut self,
fn_id: NodeId,
has_self: bool,
inputs: impl Iterator<Item = (Option<&'ast Pat>, &'ast Ty)> + Clone,
output_ty: &'ast FnRetTy
)

Perform resolution of a function signature, accounting for lifetime elision.

fn resolve_fn_params(
&mut self,
has_self: bool,
inputs: impl Iterator<Item = (Option<&'ast Pat>, &'ast Ty)>
) -> Result<LifetimeRes, (Vec<MissingLifetime>, Vec<ElisionFnParameter>)>

Resolve inside function parameters and parameter types. Returns the lifetime for elision in fn return type, or diagnostic information in case of elision failure.

fn find_lifetime_for_self(&self, ty: &'ast Ty) -> Set1<LifetimeRes>

List all the lifetimes that appear in the provided type.

fn resolve_label(
&mut self,
label: Ident
) -> Result<(NodeId, Span), ResolutionError<'a>>

Searches the current set of local scopes for labels. Returns the NodeId of the resolved label and reports an error if the label is not found or is unreachable.

fn is_label_valid_from_rib(&self, rib_index: usize) -> bool

Determine whether or not a label from the rib_indexth label rib is reachable.

fn resolve_adt(&mut self, item: &'ast Item, generics: &'ast Generics)

fn future_proof_import(&mut self, use_tree: &UseTree)

fn resolve_item(&mut self, item: &'ast Item)

fn with_generic_param_rib<'c, F>(
&'c mut self,
params: &'c [GenericParam],
kind: RibKind<'a>,
lifetime_kind: LifetimeRibKind,
f: F
)where
F: FnOnce(&mut Self),

fn with_label_rib(&mut self, kind: RibKind<'a>, f: impl FnOnce(&mut Self))

fn with_item_rib(&mut self, f: impl FnOnce(&mut Self))

fn with_constant_rib(
&mut self,
is_repeat: IsRepeatExpr,
may_use_generics: ConstantHasGenerics,
item: Option<(Ident, ConstantItemKind)>,
f: impl FnOnce(&mut Self)
)

fn with_current_self_type<T>(
&mut self,
self_type: &Ty,
f: impl FnOnce(&mut Self) -> T
) -> T

fn with_current_self_item<T>(
&mut self,
self_item: &Item,
f: impl FnOnce(&mut Self) -> T
) -> T

fn resolve_trait_items(&mut self, trait_items: &'ast [P<AssocItem>])

When evaluating a trait use its associated types’ idents for suggestions in E0412.

fn with_optional_trait_ref<T>(
&mut self,
opt_trait_ref: Option<&TraitRef>,
self_type: &'ast Ty,
f: impl FnOnce(&mut Self, Option<DefId>) -> T
) -> T

This is called to resolve a trait reference from an impl (i.e., impl Trait for Foo).

fn with_self_rib_ns(
&mut self,
ns: Namespace,
self_res: Res<NodeId>,
f: impl FnOnce(&mut Self)
)

fn with_self_rib(&mut self, self_res: Res<NodeId>, f: impl FnOnce(&mut Self))

fn resolve_implementation(
&mut self,
generics: &'ast Generics,
opt_trait_reference: &'ast Option<TraitRef>,
self_type: &'ast Ty,
item_id: NodeId,
impl_items: &'ast [P<AssocItem>]
)

fn resolve_impl_item(&mut self, item: &'ast AssocItem)

fn check_trait_item<F>(
&mut self,
id: NodeId,
ident: Ident,
kind: &AssocItemKind,
ns: Namespace,
span: Span,
err: F
)where
F: FnOnce(Ident, String, Option<Symbol>) -> ResolutionError<'a>,

fn resolve_params(&mut self, params: &'ast [Param])

fn resolve_local(&mut self, local: &'ast Local)

fn binding_mode_map(&mut self, pat: &Pat) -> FxHashMap<Ident, BindingInfo>

build a map from pattern identifiers to binding-info’s. this is done hygienically. This could arise for a macro that expands into an or-pattern where one ‘x’ was from the user and one ‘x’ came from the macro.

fn is_base_res_local(&self, nid: NodeId) -> bool

fn check_consistent_bindings(
&mut self,
pats: &[P<Pat>]
) -> Vec<FxHashMap<Ident, BindingInfo>>

Checks that all of the arms in an or-pattern have exactly the same set of bindings, with the same binding modes for each.

fn check_consistent_bindings_top(&mut self, pat: &'ast Pat)

Check the consistency of the outermost or-patterns.

fn resolve_arm(&mut self, arm: &'ast Arm)

fn resolve_pattern_top(&mut self, pat: &'ast Pat, pat_src: PatternSource)

Arising from source, resolve a top level pattern.

fn resolve_pattern(
&mut self,
pat: &'ast Pat,
pat_src: PatternSource,
bindings: &mut SmallVec<[(PatBoundCtx, FxHashSet<Ident>); 1]>
)

fn resolve_pattern_inner(
&mut self,
pat: &Pat,
pat_src: PatternSource,
bindings: &mut SmallVec<[(PatBoundCtx, FxHashSet<Ident>); 1]>
)

Resolve bindings in a pattern. This is a helper to resolve_pattern.

bindings

A stack of sets of bindings accumulated.

In each set, PatBoundCtx::Product denotes that a found binding in it should be interpreted as re-binding an already bound binding. This results in an error. Meanwhile, PatBound::Or denotes that a found binding in the set should result in reusing this binding rather than creating a fresh one.

When called at the top level, the stack must have a single element with PatBound::Product. Otherwise, pushing to the stack happens as or-patterns (p_0 | ... | p_n) are encountered and the context needs to be switched to PatBoundCtx::Or and then PatBoundCtx::Product for each p_i. When each p_i has been dealt with, the top set is merged with its parent. When a whole or-pattern has been dealt with, the thing happens.

See the implementation and fresh_binding for more details.

fn fresh_binding(
&mut self,
ident: Ident,
pat_id: NodeId,
pat_src: PatternSource,
bindings: &mut SmallVec<[(PatBoundCtx, FxHashSet<Ident>); 1]>
) -> Res<NodeId>

fn innermost_rib_bindings(
&mut self,
ns: Namespace
) -> &mut FxHashMap<Ident, Res<NodeId>>

fn try_resolve_as_non_binding(
&mut self,
pat_src: PatternSource,
ann: BindingAnnotation,
ident: Ident,
has_sub: bool
) -> Option<Res<NodeId>>

fn smart_resolve_path(
&mut self,
id: NodeId,
qself: Option<&QSelf>,
path: &Path,
source: PathSource<'ast>
)

fn smart_resolve_path_fragment(
&mut self,
qself: Option<&QSelf>,
path: &[Segment],
source: PathSource<'ast>,
finalize: Finalize
) -> PartialRes

fn self_type_is_available(&mut self) -> bool

fn self_value_is_available(&mut self, self_span: Span) -> bool

fn report_error(&mut self, span: Span, resolution_error: ResolutionError<'a>)

A wrapper around Resolver::report_error.

This doesn’t emit errors for function bodies if this is rustdoc.

fn should_report_errs(&self) -> bool

If we’re actually rustdoc then avoid giving a name resolution error for cfg() items.

fn resolve_qpath_anywhere(
&mut self,
qself: Option<&QSelf>,
path: &[Segment],
primary_ns: Namespace,
span: Span,
defer_to_typeck: bool,
finalize: Finalize
) -> Result<Option<PartialRes>, Spanned<ResolutionError<'a>>>

fn resolve_qpath(
&mut self,
qself: Option<&QSelf>,
path: &[Segment],
ns: Namespace,
finalize: Finalize
) -> Result<Option<PartialRes>, Spanned<ResolutionError<'a>>>

Handles paths that may refer to associated items.

fn with_resolved_label(
&mut self,
label: Option<Label>,
id: NodeId,
f: impl FnOnce(&mut Self)
)

fn resolve_labeled_block(
&mut self,
label: Option<Label>,
id: NodeId,
block: &'ast Block
)

fn resolve_block(&mut self, block: &'ast Block)

fn resolve_anon_const(
&mut self,
constant: &'ast AnonConst,
is_repeat: IsRepeatExpr
)

fn resolve_inline_const(&mut self, constant: &'ast AnonConst)

fn resolve_expr(&mut self, expr: &'ast Expr, parent: Option<&'ast Expr>)

fn record_candidate_traits_for_expr_if_necessary(&mut self, expr: &'ast Expr)

fn traits_in_scope(&mut self, ident: Ident, ns: Namespace) -> Vec<TraitCandidate>