Type Alias rustc_resolve::late::BindingMap

type BindingMap = FxHashMap<Ident, BindingInfo>;

Map from the name in a pattern to its binding mode.

Aliased Type

struct BindingMap {
    base: HashMap<Ident, BindingInfo, BuildHasherDefault<FxHasher>, Global>,
}

Fields
base: HashMap<Ident, BindingInfo, BuildHasherDefault<FxHasher>, Global>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference’s “Type Layout” chapter for details on type layout guarantees.

Size:32 bytes