rustc_resolve::rustdoc

Function add_doc_fragment

pub fn add_doc_fragment(out: &mut String, frag: &DocFragment)

The goal of this function is to apply the DocFragment transformation that is required when transforming into the final Markdown, which is applying the computed indent to each line in each doc fragment (a DocFragment can contain multiple lines in case of #[doc = ""]).

Note: remove the trailing newline where appropriate