rustc_resolve::rustdoc
Function prepare_to_doc_link_resolution

pub fn prepare_to_doc_link_resolution(
    doc_fragments: &[DocFragment],
) -> FxIndexMap<Option<DefId>, String>

Return the doc-comments on this item, grouped by the module they came from. The module can be different if this is a re-export with added documentation.

The last newline is not trimmed so the produced strings are reusable between early and late doc link resolution regardless of their position.