Trait rustc_resolve::ToNameBinding

pub trait ToNameBinding<'a> {
    fn to_name_binding(
        self, 
        arenas: &'a ResolverArenas<'a>
    ) -> &'a NameBinding<'a>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn to_name_binding(self, arenas: &'a ResolverArenas<'a>) -> &'a NameBinding<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?