rustc_resolve

Type Alias CmResolver

pub(crate) type CmResolver<'r, 'ra, 'tcx> = RefOrMut<'r, Resolver<'ra, 'tcx>>;

A wrapper around &mut Resolver that may be mutable or immutable, depending on a conditions.

Cm stands for “conditionally mutable”.

Prefer constructing it through Resolver::cm to ensure correctness.

Aliased Type

pub(crate) struct CmResolver<'r, 'ra, 'tcx> {
    p: &'r mut Resolver<'ra, 'tcx>,
    mutable: bool,
}

Fields
p: &'r mut Resolver<'ra, 'tcx>mutable: bool
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 16 bytes