Enum rustc_resolve::ResolutionError

pub enum ResolutionError<'a> {
    TypeParametersFromOuterFunction,
    OuterTypeParameterContext,
    NameAlreadyUsedInTypeParameterList(Name),
    IsNotATrait(&'a str),
    UndeclaredTraitName(&'a str),
    UndeclaredAssociatedType,
    MethodNotMemberOfTrait(Name, &'a str),
    TypeNotMemberOfTrait(Name, &'a str),
    ConstNotMemberOfTrait(Name, &'a str),
    VariableNotBoundInPattern(Name, usize),
    VariableBoundWithDifferentMode(Name, usize),
    VariableNotBoundInParentPattern(Name, usize),
    SelfUsedOutsideImplOrTrait,
    UseOfUndeclared(&'a str, &'a str),
    DeclarationShadowsEnumVariantOrUnitLikeStruct(Name),
    OnlyIrrefutablePatternsAllowedHere(DefId, Name),
    IdentifierBoundMoreThanOnceInParameterList(&'a str),
    IdentifierBoundMoreThanOnceInSamePattern(&'a str),
    StaticVariableReference,
    NotAnEnumVariantStructOrConst(&'a str),
    UnresolvedEnumVariantStructOrConst(&'a str),
    NotAnAssociatedConst(&'a str),
    UnresolvedAssociatedConst(&'a str),
    DoesNotNameAStruct(&'a str),
    StructVariantUsedAsFunction(&'a str),
    SelfNotAvailableInStaticMethod,
    UnresolvedName(&'a str, &'a str),
    UndeclaredLabel(&'a str),
    CannotUseRefBindingModeWith(&'a str),
    DuplicateDefinition(&'a str, Name),
    SelfImportsOnlyAllowedWithin,
    SelfImportCanOnlyAppearOnceInTheList,
    SelfImportOnlyInImportListWithNonEmptyPrefix,
    UnresolvedImport(Option<(&'a str, &'a str)>),
    FailedToResolve(&'a str),
    CannotCaptureDynamicEnvironmentInFnItem,
    AttemptToUseNonConstantValueInConstant,
}

Unstable (rustc_private)

Variants

TypeParametersFromOuterFunction
Unstable (rustc_private)

error E0401: can't use type parameters from outer function

OuterTypeParameterContext
Unstable (rustc_private)

error E0402: cannot use an outer type parameter in this context

NameAlreadyUsedInTypeParameterList
Unstable (rustc_private)

error E0403: the name is already used for a type parameter in this type parameter list

IsNotATrait
Unstable (rustc_private)

error E0404: is not a trait

UndeclaredTraitName
Unstable (rustc_private)

error E0405: use of undeclared trait name

UndeclaredAssociatedType
Unstable (rustc_private)

error E0406: undeclared associated type

MethodNotMemberOfTrait
Unstable (rustc_private)

error E0407: method is not a member of trait

TypeNotMemberOfTrait
Unstable (rustc_private)

error E0437: type is not a member of trait

ConstNotMemberOfTrait
Unstable (rustc_private)

error E0438: const is not a member of trait

VariableNotBoundInPattern
Unstable (rustc_private)

error E0408: variable {} from pattern #1 is not bound in pattern

VariableBoundWithDifferentMode
Unstable (rustc_private)

error E0409: variable is bound with different mode in pattern #{} than in pattern #1

VariableNotBoundInParentPattern
Unstable (rustc_private)

error E0410: variable from pattern is not bound in pattern #1

SelfUsedOutsideImplOrTrait
Unstable (rustc_private)

error E0411: use of Self outside of an impl or trait

UseOfUndeclared
Unstable (rustc_private)

error E0412: use of undeclared

DeclarationShadowsEnumVariantOrUnitLikeStruct
Unstable (rustc_private)

error E0413: declaration shadows an enum variant or unit-like struct in scope

OnlyIrrefutablePatternsAllowedHere
Unstable (rustc_private)

error E0414: only irrefutable patterns allowed here

IdentifierBoundMoreThanOnceInParameterList
Unstable (rustc_private)

error E0415: identifier is bound more than once in this parameter list

IdentifierBoundMoreThanOnceInSamePattern
Unstable (rustc_private)

error E0416: identifier is bound more than once in the same pattern

StaticVariableReference
Unstable (rustc_private)

error E0417: static variables cannot be referenced in a pattern

NotAnEnumVariantStructOrConst
Unstable (rustc_private)

error E0418: is not an enum variant, struct or const

UnresolvedEnumVariantStructOrConst
Unstable (rustc_private)

error E0419: unresolved enum variant, struct or const

NotAnAssociatedConst
Unstable (rustc_private)

error E0420: is not an associated const

UnresolvedAssociatedConst
Unstable (rustc_private)

error E0421: unresolved associated const

DoesNotNameAStruct
Unstable (rustc_private)

error E0422: does not name a struct

StructVariantUsedAsFunction
Unstable (rustc_private)

error E0423: is a struct variant name, but this expression uses it like a function name

SelfNotAvailableInStaticMethod
Unstable (rustc_private)

error E0424: self is not available in a static method

UnresolvedName
Unstable (rustc_private)

error E0425: unresolved name

UndeclaredLabel
Unstable (rustc_private)

error E0426: use of undeclared label

CannotUseRefBindingModeWith
Unstable (rustc_private)

error E0427: cannot use ref binding mode with ...

DuplicateDefinition
Unstable (rustc_private)

error E0428: duplicate definition

SelfImportsOnlyAllowedWithin
Unstable (rustc_private)

error E0429: self imports are only allowed within a { } list

SelfImportCanOnlyAppearOnceInTheList
Unstable (rustc_private)

error E0430: self import can only appear once in the list

SelfImportOnlyInImportListWithNonEmptyPrefix
Unstable (rustc_private)

error E0431: self import can only appear in an import list with a non-empty prefix

UnresolvedImport
Unstable (rustc_private)

error E0432: unresolved import

FailedToResolve
Unstable (rustc_private)

error E0433: failed to resolve

CannotCaptureDynamicEnvironmentInFnItem
Unstable (rustc_private)

error E0434: can't capture dynamic environment in a fn item

AttemptToUseNonConstantValueInConstant
Unstable (rustc_private)

error E0435: attempt to use a non-constant value in a constant