Module rustc_resolve::imports

A bunch of methods and structures more or less related to resolving imports.

Structs

ImportData 🔒

One import.

NameResolution 🔒

Records information about the resolution of a name in a namespace of a module.

UnresolvedImportError 🔒

An error that may be transformed into a diagnostic later. Used to combine multiple unresolved import errors within the same use tree into a single diagnostic.

Enums

ImportKind 🔒

Contains data for specific kinds of imports.

Functions

import_kind_to_string 🔒

import_path_to_string 🔒

pub_use_of_private_extern_crate_hack 🔒

Type Aliases

Import 🔒

All imports are unique and allocated on a same arena, so we can use referential equality to compare them.

Res 🔒