rustc_resolve::late
Module diagnostics

Structs

BaseError ğŸ”’

Only used for diagnostics.

ElisionFnParameter ğŸ”’

Description of the lifetimes appearing in a function parameter. This is used to provide a literal explanation to the elision failure.

LifetimeFinder ğŸ”’

MissingLifetime ğŸ”’

Description of an elided lifetime.

Enums

AssocSuggestion ğŸ”’

A field or associated item from self type suggested in case of resolution failure.

LifetimeElisionCandidate ğŸ”’

Description of lifetimes that appear as candidates for elision. This is used to suggest introducing an explicit lifetime.

TypoCandidate ğŸ”’

Functions

import_candidate_to_enum_paths ğŸ”’

Gets the stringified path for an enum from an ImportSuggestion for an enum variant.

is_self_type ğŸ”’

is_self_value ğŸ”’

mk_where_bound_predicate ğŸ”’

signal_label_shadowing ğŸ”’

Shadowing involving a label is only a warning for historical reasons.

signal_lifetime_shadowing ğŸ”’

Report lifetime/lifetime shadowing as an error.

Type Aliases

Res ğŸ”’