Struct rustc_resolve::Resolver

pub struct Resolver<'a> {
    pub definitions: Definitions,
    pub def_map: DefMap,
    pub freevars: FreevarMap,
    pub export_map: ExportMap,
    pub trait_map: TraitMap,
    pub make_glob_map: bool,
    pub glob_map: GlobMap,
    pub maybe_unused_trait_imports: NodeSet,
    // some fields omitted
}

Unstable (rustc_private)

The main resolver class.

Fields
definitions: Definitions
Unstable (rustc_private)
def_map: DefMap
Unstable (rustc_private)
freevars: FreevarMap
Unstable (rustc_private)
export_map: ExportMap
Unstable (rustc_private)
trait_map: TraitMap
Unstable (rustc_private)
make_glob_map: bool
Unstable (rustc_private)
glob_map: GlobMap
Unstable (rustc_private)
maybe_unused_trait_imports: NodeSet
Unstable (rustc_private)

Methods

impl<'b> Resolver<'b>

fn build_reduced_graph(&mut self, krate: &Crate)

Unstable (rustc_private)

Constructs the reduced graph for the entire crate.

fn populate_module_if_necessary(&mut self, module: Module<'b>)

Unstable (rustc_private)

Ensures that the reduced graph rooted at the given external module is built, building it if it is not.

impl<'a> Resolver<'a>

fn resolve_imports(&mut self)

Unstable (rustc_private)

impl<'a> Resolver<'a>

fn resolve_name_in_module(&mut self, module: Module<'a>, name: Name, ns: Namespace, allow_private_imports: bool, record_used: Option<Span>) -> ResolveResult<&'a NameBinding<'a>>

Unstable (rustc_private)

Attempts to resolve the supplied name in the given module for the given namespace. If successful, returns the binding corresponding to the name. Invariant: if record_used is Some, import resolution must be complete.

fn add_import_directive(&mut self, module_path: Vec<Name>, subclass: ImportDirectiveSubclass<'a>, span: Span, id: NodeId, vis: Visibility)

Unstable (rustc_private)

fn try_define<T>(&mut self, module: Module<'a>, name: Name, ns: Namespace, binding: T) -> Result<(), &'a NameBinding<'a>> where T: ToNameBinding<'a>

Unstable (rustc_private)

impl<'a> Resolver<'a>

fn assign_node_ids(&mut self, krate: Crate) -> Crate

Unstable (rustc_private)

impl<'a> Resolver<'a>

fn new(session: &'a Session, make_glob_map: MakeGlobMap, arenas: &'a ResolverArenas<'a>) -> Resolver<'a>

Unstable (rustc_private)

fn arenas() -> ResolverArenas<'a>

Unstable (rustc_private)

fn resolve_crate(&mut self, krate: &Crate)

Unstable (rustc_private)

Entry point to crate resolution.

fn with_module_lexical_scope<T, F>(&mut self, module: Module<'a>, f: F) -> T where F: FnOnce(&mut Resolver<'a>) -> T

Unstable (rustc_private)